return to uint16_t

+++ dNet/AuthPackets.h
@@ -7,7 +7,7 @@
 #include "magic_enum.hpp"
 
 enum class eLoginResponse : uint8_t;
-enum class ServiceType : uint8_t;
+enum class ServiceType : uint16_t;
 class dServer;
 
 enum class eStamps : uint32_t {
